+++ CHANGELOG.md
@@ -7,6 +7,18 @@
  - Lua: Add event_to_json() function, that will convert a Slipstream event to a json string. Change debug.lua to use that, making it trivial.
  - Lua: At start: Make sure all Lua script before IRacingPlugin publishes events
  - UI: Will show all events, even if they are published before UI was ready
+ - Events renamed for consistency:
+   - CommandPlayAudio to AudioCommandPlay
+   - CommandSay to AudioCommandSay
+   - CommandPluginEnable to InternalCommandPluginEnable
+   - CommandPluginDisable to InternalCommandPluginDisable
+   - CommandPluginRegister to InternalCommandPluginRegister
+   - CommandPluginStates to InternalCommandPluginStates
+   - CommandPluginUnregister to InternalCommandPluginUnregister
+   - Initialized to InternalInitialized
+   - PluginState to InternalPluginState
+   - CommandTwitchSendMessage to TwitchCommandSendMessage
+   - CommandWriteToConsole to UICommandWriteToConsole
 
 ## [0.2.0](https://github.com/dennis/slipstream/releases/tag/v0.2.0) (2020-12-30)
 [Full Changelog](https://github.com/dennis/slipstream/compare/v0.1.0...v0.2.0)